smartcity/requirements.txt
asgiref==3.2.7
astroid==2.3.3
Babel==2.8.0
beautifulsoup4==4.9.1
colorama==0.4.3
Django==3.0.7
django-autoslug-iplweb==1.9.4.dev0
django-bootstrap4==1.1.1
django-cities-light==3.6.0
django-dotenv==1.4.2
django-phonenumber-field==4.0.0
flake8==3.8.1
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.12.4
Pillow==7.1.2
progressbar2==3.51.3
psycopg2==2.8.5
pycodestyle==2.6.0
pycountry==19.8.18
pyflakes==2.2.0
pylint==2.4.4
python-utils==2.4.0
pytz==2020.1
six==1.13.0
soupsieve==2.0.1
sqlparse==0.3.1
typed-ast==1.4.0
Unidecode==1.1.1
wrapt==1.11.2
zipp==3.1.0
